% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{DayOfBirth}
\alias{DayOfBirth}
\title{Day of Birth}
\format{
A data frame with 7 observations on the following 2 variables.
\describe{ \item{day}{a character vector} \item{births}{a
numeric vector} }
}
\source{
Ventura, S.J., J.A. Martin, S.C. Curtin, F. Menacker, and B.E.
Hamilton. 2001. Births: final data for 1999. \emph{National Vital Statistics
Reports} Vol. 49, No. 1.
}
\usage{
DayOfBirth
}
\description{
Day of the week for 350 U.S. births in 1999.
}
\references{
\url{http://cdc.gov/NCHS/products/nvsr.htm}
}
\keyword{datasets}
