% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SexualSelection}
\alias{SexualSelection}
\title{Sexual Conflict}
\format{
A data frame with 25 observations on the following 4 variables.
\describe{ \item{polyandrous.species}{a numeric vector}
\item{monandrous.species}{a numeric vector}
\item{difference}{a numeric vector}
\item{taxon.pair}{identifier} }
}
\source{
Arnqvist, G., M. Edvardsson, U. Friberg, and T. Nilsson. 2000.
Sexual conflict promotes speciation in insects. \emph{Proceedings of the
National Academy of Sciences (USA)} 97: 10460-10464.
}
\usage{
SexualSelection
}
\description{
Number of species in each of two taxa in closely related taxon pairings and
the difference between the two groups. One taxon has multiple matings
(\code{polyandrous.species}) and one has only single matings
(\code{monandrous.species}).
}
\references{
\url{http://www.ncbi.nlm.nih.gov/pmc/articles/PMC27046/}
}
\keyword{datasets}
