% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Toads}
\alias{Toads}
\title{Right-handed Toads}
\format{
A data frame with 19 observations on the following 2 variables.
\describe{ \item{n.toads}{a numeric vector} \item{prob}{a
numeric vector} }
}
\description{
Hypothetical probability of a toad being right-handed
}
\keyword{datasets}
