% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TwoKids}
\alias{TwoKids}
\title{Number of Boys in Two-Child Families}
\format{
A data frame with 3 observations on the following 2 variables.
\describe{ \item{num.boys}{a numeric vector} \item{count}{a
numeric vector} }
}
\source{
Rodgers, J.L. and D. Doughty. 2001. Does having boys or girls run in
the family? \emph{Chance Magazine} Fall 2001: 8-13.
}
\description{
The number of boys in a sample of 2,444 two-child families.
}
\references{
\url{http://www.dartmouth.edu/~chance/chance_news/recent_news/chance_news_10.11.html#item13}
}
\keyword{datasets}
